requirements.txt
python-telegram-bot[job-queue]==20.7
apscheduler
aiofiles
flask
python-dotenv
